The new "Read only" button changes
The read only button now turns the background image on and off on the
homepage as well! It also changes the colour of the title and
description to grey so it can still be seen. Jayden will make a new
hamburger icon for this soon so don’t worry that they hamburger goes
invisible when read only is turned on.